bug, sometime selling rate round to int.

+++ models/CurrencyExchangeOrder.php
@@ -435,7 +435,7 @@ class CurrencyExchangeOrder extends ActiveRecord
                 (
                     isset($changedAttributes['selling_rate']) &&
                     $this->selling_rate != '' &&
-                    ($this->selling_rate = intval($this->selling_rate)) &&
+                    ($this->selling_rate = floatval($this->selling_rate)) &&
                     $this->selling_rate != $changedAttributes['selling_rate']
                 ) || $insert
             )
@@ -454,7 +454,7 @@ class CurrencyExchangeOrder extends ActiveRecord
                 (
                     isset($changedAttributes['buying_rate']) &&
                     $this->selling_rate != '' &&
-                    ($this->buying_rate = intval($this->buying_rate)) &&
+                    ($this->buying_rate = floatval($this->buying_rate)) &&
                     $this->buying_rate != $changedAttributes['buying_rate']
                 ) || $insert
             )
